+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Info screen: add post status to all products
+
 ### 1.2.0: July 29th, 2020
 * Info screen now displays draft products
 * Update FR translation